neovim: extract lsp locate helper with progress feedback
