+++ .config/nvim/lua/rc/keymaps/plugins.lua
@@ -1,89 +1,8 @@
 local set = require("rc.keymaps.util").set
-
-local function snake_case(value)
-  return value:gsub("::", "/"):gsub("([A-Z]+)([A-Z][a-z])", "%1_%2"):gsub("([a-z%d])([A-Z])", "%1_%2"):lower()
-end
-
-local function lsp_location_uri(location)
-  return location.uri or location.targetUri
-end
-
-local function lsp_location_path(location)
-  local uri = lsp_location_uri(location)
-  return uri and vim.uri_to_fname(uri) or nil
-end
-
-local function best_lsp_location(result)
-  if not result then
-    return nil
-  end
-  if result.uri or result.targetUri then
-    return result
-  end
-
-  local symbol_path = snake_case(vim.fn.expand("<cword>")) .. ".rb"
-  for _, location in ipairs(result) do
-    local path = lsp_location_path(location)
-    if path and path:sub(-#symbol_path) == symbol_path then
-      return location
-    end
-  end
-
-  return result[1]
-end
-
-local function jump_to_lsp_location(location, client_id)
-  local client = vim.lsp.get_client_by_id(client_id)
-  local offset_encoding = client and client.offset_encoding or "utf-16"
-
-  if location.targetUri then
-    location = {
-      uri = location.targetUri,
-      range = location.targetSelectionRange or location.targetRange,
-    }
-  end
-
-  vim.lsp.util.jump_to_location(location, offset_encoding, true)
-end
+local lsp_location = require("rc.lsp.locate").location
 
 local function has_deno()
   return vim.fn.executable("deno") == 1
-end
-
-local function lsp_location(method, extend_params)
-  return function()
-    local clients = vim.tbl_filter(function(client)
-      return client:supports_method(method, 0)
-    end, vim.lsp.get_clients({ bufnr = 0 }))
-    local client = clients[1]
-    if not client then
-      return
-    end
-
-    local offset_encoding = client and client.offset_encoding or "utf-16"
-    local params = vim.lsp.util.make_position_params(0, offset_encoding)
-    local jumped = false
-
-    if extend_params then
-      extend_params(params)
-    end
-
-    for _, lsp_client in ipairs(clients) do
-      lsp_client:request(method, params, function(err, result, ctx)
-        if jumped or err then
-          return
-        end
-
-        local location = best_lsp_location(result)
-        if not location then
-          return
-        end
-
-        jumped = true
-        jump_to_lsp_location(location, ctx.client_id)
-      end, 0)
-    end
-  end
 end
 
 return function()
